Test calucaltions for bearing
